# backend/openapi-schema.yaml
openapi: 3.0.3
info:
  title: VYBE Music API
  version: v1
  description: Gen-Z Music Discovery Engine REST API
paths:
...
  /api/v1/home/:
    get:
      operationId: home_feed
      description: Composed home payload containing greeting, hero banner, trending
        tracks, mood crates, recently played history, quick picks, and playlists.
      summary: Get Home Dashboard Feed
      tags:
      - home
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HomeFeedResponse'
          description: ''
...
  /api/v1/music/moods/:
    get:
      operationId: music_moods_list
      description: Retrieve all available Gen-Z mood categories with illustrations
        and accent colors.
      summary: Get Moods
      tags:
      - music
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Mood'
          description: ''
  /api/v1/music/moods/{slug}/songs/:
    get:
      operationId: music_mood_songs
      description: Retrieve curated tracks belonging to a specific mood category.
      summary: Get Mood Tracks
      parameters:
      - in: query
        name: limit
        schema:
          type: integer
          default: 20
        description: 'Maximum number of tracks to return (default: 20)'
      - in: path
        name: slug
        schema:
          type: string
        required: true
      tags:
      - music
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MoodDetailResponse'
          description: ''
        '404':
          description: Mood not found
  /api/v1/music/recommendations/:
    get:
      operationId: music_recommendations_list
      description: Generates heuristic-scored music recommendations based on user
        history, liked songs, preferred moods, or seed track/mood.
      summary: Get Personalized Recommendations
      parameters:
      - in: query
        name: limit
        schema:
          type: integer
          default: 20
        description: 'Maximum number of recommendations (default: 20)'
      - in: query
        name: mood
        schema:
          type: string
        description: Optional mood filter/boost (e.g. chill, lockin, mainchar)
      - in: query
        name: seed_song_id
        schema:
          type: string
        description: Optional seed YouTube video ID to base recommendations around
      tags:
      - music
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Song'
          description: ''
...
  /api/v1/music/songs/{id}/like/:
    post:
      operationId: music_like_song
      description: Like a song by YouTube video ID or database UUID.
      summary: Like Song
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - music
      security:
      - jwtAuth: []
      responses:
        '200':
          description: Song liked
    delete:
      operationId: music_unlike_song
      description: Remove song from listener's liked songs.
      summary: Unlike Song
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - music
      security:
      - jwtAuth: []
      responses:
        '200':
          description: Song unliked
  /api/v1/music/songs/{id}/similar/:
    get:
      operationId: music_song_similar
      description: Retrieve similar tracks and up next queue for a given YouTube video
        ID.
      summary: Get Similar Tracks (Track Radio)
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      - in: query
        name: limit
        schema:
          type: integer
          default: 20
        description: 'Maximum number of similar tracks (default: 20)'
      tags:
      - music
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Song'
          description: ''
  /api/v1/music/trending/:
    get:
      operationId: music_trending_list
      description: Retrieve curated and top trending tracks with Redis caching.
      summary: Get Trending Tracks
      parameters:
      - in: query
        name: limit
        schema:
          type: integer
          default: 20
        description: 'Maximum number of tracks to return (default: 20)'
      tags:
      - music
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Song'
          description: ''
...
components:
  schemas:
...
    AlbumSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        providerId:
          type: string
        title:
          type: string
          maxLength: 255
        thumbnailUrl:
          type: string
          format: uri
        year:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
          nullable: true
      required:
      - id
      - providerId
      - thumbnailUrl
      - title
    ArtistSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        providerId:
          type: string
        name:
          type: string
          maxLength: 255
        thumbnailUrl:
          type: string
          format: uri
        avatarColor:
          type: string
      required:
      - avatarColor
      - id
      - name
      - providerId
      - thumbnailUrl
...
    Genre:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 100
        slug:
          type: string
          maxLength: 100
          pattern: ^[-a-zA-Z0-9_]+$
      required:
      - id
      - name
      - slug
    HomeBanner:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        subtitle:
          type: string
        accentColor:
          type: string
        illustration:
          type: string
        badge:
          type: string
        actionType:
          type: string
        actionTarget:
          type: string
      required:
      - accentColor
      - actionTarget
      - actionType
      - badge
      - id
      - illustration
      - subtitle
      - title
    HomeFeedResponse:
      type: object
      properties:
        greeting:
          type: string
        user:
          type: object
          additionalProperties: {}
          nullable: true
        banner:
          $ref: '#/components/schemas/HomeBanner'
        trending:
          type: array
          items:
            $ref: '#/components/schemas/Song'
        moods:
          type: array
          items:
            $ref: '#/components/schemas/Mood'
        recentlyPlayed:
          type: array
          items:
            $ref: '#/components/schemas/Song'
        quickPicks:
          type: array
          items:
            $ref: '#/components/schemas/Song'
        playlists:
          type: array
          items: {}
      required:
      - banner
      - greeting
      - moods
      - playlists
      - quickPicks
      - recentlyPlayed
      - trending
      - user
    IllustrationEnum:
      enum:
      - chill
      - lockin
      - mainchar
      - unhinged
      - vinyl
      - cassette
      - lofi
      - retro
      type: string
      description: |-
        * `chill` - Chill
        * `lockin` - Lock In
        * `mainchar` - Main Character
        * `unhinged` - Unhinged
        * `vinyl` - Vinyl
        * `cassette` - Cassette
        * `lofi` - Lo-Fi
        * `retro` - Retro
...
    Mood:
      type: object
      properties:
        slug:
          type: string
          maxLength: 50
          pattern: ^[-a-zA-Z0-9_]+$
        title:
          type: string
          maxLength: 100
        subtitle:
          type: string
          maxLength: 200
        accentColor:
          type: string
        bgClass:
          type: string
        rotation:
          type: string
          maxLength: 50
        description:
          type: string
        illustration:
          type: string
          maxLength: 50
      required:
      - accentColor
      - bgClass
      - slug
      - title
    MoodDetailResponse:
      type: object
      properties:
        mood:
          $ref: '#/components/schemas/Mood'
        tracks:
          type: array
          items:
            $ref: '#/components/schemas/Song'
      required:
      - mood
      - tracks
...
    Song:
      type: object
      description: |-
        Normalized Song contract:
        {
            "id": "...",
            "youtube_video_id": "...",
            "title": "...",
            "artist": { "id": "...", "name": "..." },
            "album": { "id": "...", "name": "..." },
            "thumbnail_url": "...",
            "duration_seconds": 210,
            ...
        }
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        youtube_video_id:
          type: string
        youtubeVideoId:
          type: string
        title:
          type: string
          maxLength: 255
        artist:
          allOf:
          - $ref: '#/components/schemas/ArtistSummary'
          readOnly: true
        album:
          allOf:
          - $ref: '#/components/schemas/AlbumSummary'
          readOnly: true
        thumbnail_url:
          type: string
          format: uri
        thumbnailUrl:
          type: string
          format: uri
        duration_seconds:
          type: integer
        durationSeconds:
          type: integer
        durationFormatted:
          type: string
          readOnly: true
        accentColor:
          type: string
        illustration:
          $ref: '#/components/schemas/IllustrationEnum'
        bpm:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
          nullable: true
        playCount:
          type: integer
          readOnly: true
        playCountFormatted:
          type: string
          readOnly: true
        genres:
          type: array
          items:
            $ref: '#/components/schemas/Genre'
          readOnly: true
        moods:
          type: array
          items:
            $ref: '#/components/schemas/Mood'
          readOnly: true
        isLiked:
          type: boolean
          readOnly: true
      required:
      - accentColor
      - album
      - artist
      - durationFormatted
      - durationSeconds
      - duration_seconds
      - genres
      - id
      - isLiked
      - moods
      - playCount
      - playCountFormatted
      - thumbnailUrl
      - thumbnail_url
      - title
      - youtubeVideoId
      - youtube_video_id
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT